allow pgmetric to be used as library
pgmetrics is a great command-line tool. but running this specific
commandline in every environment is not ideal.

making `collect` public, it can be used as library.

pgmetrics.Model is already available to other packages.

this is the minimal change one can make in this to make it work like a
library. separating it out as a real library can be taken as a separate
task and i think it would be only necessary if this simple setup doesn't
work.